removed a million not needed files, fixed all errors and warings, unused vars mainly

+++ src/signal.c
@@ -6,7 +6,7 @@
 /*   By: joonasmykkanen <joonasmykkanen@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/12 11:07:49 by joonasmykka       #+#    #+#             */
-/*   Updated: 2023/05/30 16:31:45 by joonasmykka      ###   ########.fr       */
+/*   Updated: 2023/05/30 17:10:30 by joonasmykka      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,22 +17,25 @@ extern t_data	g_data;
 
 void	handle_ctrl_c(int sig)
 {
-	const char nl[2] = {10, 0};
+	const char nlc[2] = {10, 0};
 	const char eof[2] = {4, 0};
 	
-	if (g_data.cur.mode_heredoc == 1)
+	if (sig == SIGINT)
 	{
-		g_data.cur.mode_heredoc = 0;
-		g_data.cur.err_flag = 1;
-		ioctl(1, TIOCSTI, eof);
+		if (g_data.cur.heredoc_mode == 1)
+		{
+			g_data.cur.heredoc_mode = 0;
+			g_data.cur.err_flag = 1;
+			ioctl(1, TIOCSTI, eof);
+		}
+		if (g_data.sig.exec_pid == NO_CHILDS)
+		{
+			ioctl(1, TIOCSTI, nlc);
+			rl_on_new_line();
+		}
+		else
+			kill(g_data.sig.exec_pid, SIGINT);
 	}
-	if (g_data.sig.exec_pid == NO_CHILDS)
-	{
-		ioctl(1, TIOCSTI, "\n");
-		rl_on_new_line();
-	}
-	else
-		kill(g_data.sig.exec_pid, SIGINT);
 }
 
 void	signal_manager(void)
